requirements.txt
torch
torchvision
continuum
scikit-learn
faiss-cpu